requirements.txt
matplotlib
scikit-image
scikit-learn
scipy
numpy
torch
torchvision
opencv-python
opencv-contrib-python
runway-python
gdown
pillow
